requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python>=4.4.0
Pillow>=8.0.0
tqdm>=4.50.0
scikit-image>=0.17.2
scipy>=1.5.0
PyQt5>=5.15.0
h5py>=2.10.0
pandas>=1.1.0
scikit-learn>=0.23.2
argparse>=1.4.0